+++ src/router/RORouteHandler.cpp
@@ -35,6 +35,7 @@
 #include <utils/xml/SUMOSAXHandler.h>
 #include <utils/xml/SUMOXMLDefinitions.h>
 #include <utils/geom/GeoConvHelper.h>
+#include <utils/common/FileHelpers.h>
 #include <utils/common/MsgHandler.h>
 #include <utils/common/StringTokenizer.h>
 #include <utils/common/UtilExceptions.h>
@@ -228,11 +229,12 @@ RORouteHandler::myStartElement(int element,
             double arrivalPos = attrs.getOpt<double>(SUMO_ATTR_ARRIVALPOS, myVehicleParameter->id.c_str(), ok,
                                 stop == nullptr ? -NUMERICAL_EPS : stop->endPos);
             const std::string desc = attrs.get<std::string>(SUMO_ATTR_LINES, pid.c_str(), ok);
-            myActivePerson->addRide(from, to, desc, arrivalPos, busStopID);
+            const std::string group = attrs.getOpt<std::string>(SUMO_ATTR_GROUP, pid.c_str(), ok, "");
+            myActivePerson->addRide(from, to, desc, arrivalPos, busStopID, group);
             break;
         }
         case SUMO_TAG_CONTAINER:
-            myActiveContainerPlan = new OutputDevice_String(false, 1);
+            myActiveContainerPlan = new OutputDevice_String(1);
             myActiveContainerPlanSize = 0;
             myActiveContainerPlan->openTag(SUMO_TAG_CONTAINER);
             (*myActiveContainerPlan) << attrs;
@@ -806,21 +808,14 @@ RORouteHandler::addTranship(const SUMOSAXAttributes& /*attrs*/) {
 void
 RORouteHandler::parseEdges(const std::string& desc, ConstROEdgeVector& into,
                            const std::string& rid, bool& ok) {
-    if (desc[0] == BinaryFormatter::BF_ROUTE) {
-        std::istringstream in(desc, std::ios::binary);
-        char c;
-        in >> c;
-        FileHelpers::readEdgeVector(in, into, rid);
-    } else {
-        for (StringTokenizer st(desc); st.hasNext();) {
-            const std::string id = st.next();
-            const ROEdge* edge = myNet.getEdge(id);
-            if (edge == nullptr) {
-                myErrorOutput->inform("The edge '" + id + "' within the route " + rid + " is not known.");
-                ok = false;
-            } else {
-                into.push_back(edge);
-            }
+    for (StringTokenizer st(desc); st.hasNext();) {
+        const std::string id = st.next();
+        const ROEdge* edge = myNet.getEdge(id);
+        if (edge == nullptr) {
+            myErrorOutput->inform("The edge '" + id + "' within the route " + rid + " is not known.");
+            ok = false;
+        } else {
+            into.push_back(edge);
         }
     }
 }
@@ -1015,6 +1010,7 @@ RORouteHandler::addPersonTrip(const SUMOSAXAttributes& attrs) {
     parseWalkPositions(attrs, myVehicleParameter->id, from, to, departPos, arrivalPos, busStopID, lastStage, ok);
 
     const std::string modes = attrs.getOpt<std::string>(SUMO_ATTR_MODES, id, ok, "");
+    const std::string group = attrs.getOpt<std::string>(SUMO_ATTR_GROUP, id, ok, "");
     SVCPermissions modeSet = 0;
     for (StringTokenizer st(modes); st.hasNext();) {
         const std::string mode = st.next();
@@ -1033,7 +1029,7 @@ RORouteHandler::addPersonTrip(const SUMOSAXAttributes& attrs) {
     const std::string types = attrs.getOpt<std::string>(SUMO_ATTR_VTYPES, id, ok, "");
     double walkFactor = attrs.getOpt<double>(SUMO_ATTR_WALKFACTOR, id, ok, OptionsCont::getOptions().getFloat("persontrip.walkfactor"));
     if (ok) {
-        myActivePerson->addTrip(from, to, modeSet, types, departPos, arrivalPos, busStopID, walkFactor);
+        myActivePerson->addTrip(from, to, modeSet, types, departPos, arrivalPos, busStopID, walkFactor, group);
     }
 }