+++ src/utils/xml/SAXWeightsHandler.h
@@ -93,6 +93,15 @@ public:
          */
         virtual void addEdgeWeight(const std::string& id, double val, double beg, double end) const = 0;
 
+        virtual void addEdgeRelWeight(const std::string& from, const std::string& to,
+                                   double val, double beg, double end) const {
+            UNUSED_PARAMETER(from);
+            UNUSED_PARAMETER(to);
+            UNUSED_PARAMETER(val);
+            UNUSED_PARAMETER(beg);
+            UNUSED_PARAMETER(end);
+        }
+
     private:
     /// @brief we made the assignment operator invalid
         EdgeFloatTimeLineRetriever& operator=(const EdgeFloatTimeLineRetriever&) = delete;
@@ -180,10 +189,14 @@ protected:
     //@}
 
 private:
-    /// @brief Parses the efforts of a lane for the previously read times
+    /// @brief Parses the data of an edge or lane for the previously read times
     void tryParse(const SUMOSAXAttributes& attrs, bool isEdge);
 
-    /// @brief List of definitions what shall be read and whereto stored while parsing the file
+    /// @brief Parses the data of an edgeRel for the previously read times
+    void tryParseEdgeRel(const SUMOSAXAttributes& attrs);
+
+private:
+    /// List of definitions what shall be read and whereto stored while parsing the file
     std::vector<ToRetrieveDefinition*> myDefinitions;
 
     /// @brief the begin of the time period that is currently processed