+++ docs/web/docs/ChangeLog.md
@@ -8,19 +8,35 @@ permalink: /ChangeLog/
 ### Bugfixes
 - Simulation
   - Fixed sublane model bug that reduced flow on intersections and caused emergency braking. Issue #6649
+  - Fixed sublane model bug that was causing cyclical switching between keep-right and sublane-alignment choices. Issue #6084
 - NETCONVERT
   - Fixed precision when filtering network with a given boundary. Issue #6206
 - SUMO-GUI
   - fixed layout of meso edge parameter dialog at high occupancy (regression in 1.5.0)
+- DUAROUTER
+  - Attribute `group` of `<ride>` and `<personTrip>` is no longer lost. Issue #6555
 
 ### Enhancements
 - Simulation
   - Actuated traffic lights now support the use of [custom detectors](Simulation/Traffic_Lights.md#custom_detectors). Issue #5125
+  - The new route attriubte `period` is now applied repeatedly when rerouting the vehicle and the new route contains stops with attribute `until` to shift those times forward (e.g. on cyclical rerouters for public transport vehicles). Issue #6550
+  - The new route attribute `repeat` can now be used to define repeating routes. When the route contains stop definitions with the `until`-attribute, the new route attribute `period` must be used to shift the repeated stop times. Issue #6554  
+- SUMO-GUI
+  - Missing data values (color by edgeData) can now be distinguished from negative data values. Issue #6660
+  - `<edgeRelation>`-data files can now be visualized. Issue #6659
+- NETCONVERT
+  - Edge attribute `spreadType` now supports value `roadCenter` which aligns edges with common geometry so that the geometry reflects the middle of the road regardless of differences in lane numbers (causing natural alignment for turning lanes). Issue #1758
+  - Added option **--default.spreadtype** to set the default spread type for edges. Issue #6662
 - Tools
   - added osmWebWizard option to import a simplified network that is only for cars. Issue #6595
+  - [routeSampler.py](Tools/Turns.md#routesampler.py) improvements
+    - supports optimization by linear programming using option **--optimize** (requires scipy)
+    - supports output of mismatch between measured counts and sampled counts for visualization in SUMO-GUI via `edgeData` and `edgeRelation` files using option **--mismatch-output**
+    - supports limited optimization of route input (without resampling) by setting option **--optimize-input**
 
 ### Other
-
+- Network version is now 1.6 (Issue #1758)
+- The turn-file format used by JTRROUTER is now deprecated. Instead the new edgeRelation-file format should be used. To convert old files into the new format, the tool turnFile2EdgeRelations.py can be used. The tools jtcrouter.py and routeSampler.py will only accept the new file format as input.
 
 ## Version 1.5.0 (11.02.2020)